requirements.txt
deepgram-sdk==3.7.7
openai==1.59.3
python-dotenv          ==  1.0.1
silero-vad             ==  5.1.2
onnxruntime            ==  1.20.1
requests               ==  2.32.3
termcolor              ==  2.5.0
aiortc                 ==  1.9.0
aiodns                 ==  3.2.0

# dev
pipdeptree==2.24.0

